pyproject.toml
[project]
name = "pomodoro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "cryptography==40.0.2",
    "fastapi>=0.115.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "ruff>=0.8.6",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = "."
